+++ semcore/core/src/utils/ref.ts
@@ -1,5 +1,4 @@
-import React, { type MutableRefObject, type Ref, type RefObject } from 'react';
-import { findDOMNode } from 'react-dom';
+import React, { type MutableRefObject, type Ref } from 'react';
 
 export function setRef<T>(ref: Ref<T>, value: T) {
   if (typeof ref === 'function') {