+++ semcore/feature-highlight/src/components/notice/Notice.tsx
@@ -84,5 +84,4 @@ export const NoticeFH = createComponent(NoticeFHRoot, {
   Title: Notice.Title,
   Text: Notice.Text,
   Close: Notice.Close,
-  CloseIcon: Notice.CloseIcon,
 }) as HighlightedNoticeComponent;