+++ stories/components/base-components/ellipsis/tests/examples/trim_with_special_text_size.tsx
@@ -1,0 +1,35 @@
+import type { EllipsisSettings, BoxProps } from '@semcore/ui/base-components';
+import type { TextProps } from '@semcore/ui/typography';
+import { Text } from '@semcore/ui/typography';
+import React from 'react';
+
+type TextExampleProps = {
+  ellipsis?: true | EllipsisSettings;
+} & TextProps & BoxProps;
+
+const Demo = (props: TextExampleProps) => {
+  return (
+    <Text
+      display='block'
+      w={props.w}
+      mb={3}
+      size={props.size}
+      tag='p'
+      mt={0}
+      style={{ outline: '1px solid red' }}
+      ellipsis={props.ellipsis}
+    >
+      WordWord WordWord
+    </Text>
+  );
+};
+
+export const defaultProps: TextExampleProps = {
+  ellipsis: true,
+  size: 400,
+  w: 100,
+};
+
+Demo.defaultProps = defaultProps;
+
+export default Demo;