+++ stories/patterns/filters/filter-search/docs/examples/dynamic_search.tsx
@@ -3,7 +3,6 @@ import Search from '@semcore/icon/Search/m';
 import { Flex } from '@semcore/ui/base-components';
 import { ButtonLink } from '@semcore/ui/button';
 import Input from '@semcore/ui/input';
-import { Hint } from '@semcore/ui/tooltip';
 import { Text } from '@semcore/ui/typography';
 import React from 'react';
 
@@ -35,7 +34,7 @@ const Demo = () => {
         />
         {value && (
           <Input.Addon>
-            <ButtonLink use='secondary' addonLeft={CloseM} title='Clear' onClick={handleClick} />
+            <ButtonLink use='secondary' addonLeft={CloseM} aria-label='Clear' onClick={handleClick} />
           </Input.Addon>
         )}
       </Input>